bus2d/requirements.txt
# BUS-2D pipeline dependencies
# Install: pip install -r requirements.txt
pyyaml>=6.0
numpy>=1.24
nibabel>=5.0
SimpleITK>=2.3
scipy>=1.10
opencv-python>=4.8
openpyxl>=3.1
matplotlib>=3.7
ultralytics>=8.1
webdataset>=0.2
